Fix how we print to terimal Because we were printing JSON to the terminal and the quotation marks were causing issues I chose to escape them by converting them to single quotes then un-escaping them when I read the line back in.
This of course caused issues, especially when reading JSON which had both single and double quotation
characters.

To fix that this change set updates the terminal writting function to properly escape the quotation
characters in a more standard way. This allows the tool to work when the JSON has both quotation
characters and is more compatible with other tools.

I also added new integration tests which excersies the CLI to verify that those commands are working
and play nice with xargs.